requirements.txt
appdirs==1.4.4
asttokens==2.2.1
backcall==0.2.0
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.2.1
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
cytoolz==0.12.0
dask==2023.4.1
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.9.0
fonttools==4.25.0
fsspec==2023.4.0
gmpy2==2.1.2
idna==3.4
imagecodecs==2021.8.26
imageio==2.26.0
importlib-metadata==6.0.0
ipdb==0.13.13
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.1
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
munkres==1.1.4
networkx==2.8.4
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
openTSNE==1.0.0
packaging==23.0
pandas==2.0.2
parso==0.8.3
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
ply==3.11
pooch==1.4.0
prompt-toolkit==3.0.39
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.29.0
roc-utils==0.2.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
setuptools==66.0.0
sip==6.6.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tensorboardX==2.2
threadpoolctl==2.2.0
tifffile==2021.7.2
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
wheel==0.38.4
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.11.0
